Moved what was in srp.go into util.go and srp_type.go into srp.go
